web_app/backend/requirements.txt
# Requirements for WALL-E Wildlife Robot Control Panel Backend
# Generated by Claude

fastapi>=0.104.0
uvicorn[standard]>=0.24.0
websockets>=12.0
opencv-python>=4.8.0
PyYAML>=6.0
python-multipart>=0.0.6
httpx>=0.25.0
aiofiles>=23.2.1
